requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-filter
environ
psycopg2-binary